+++ assembly/instructions/bit.ts
@@ -9,7 +9,7 @@ export const count_set_bits_64: InstructionRun = (r, args, regs) => {
 
 // COUNT_SET_BITS_32
 export const count_set_bits_32: InstructionRun = (r, args, regs) => {
-  regs[reg(args.b)] = popcnt<u32>(u32(regs[reg(args.a)]));
+  regs[reg(args.b)] = u64(popcnt<u32>(u32(regs[reg(args.a)])));
   return ok(r);
 };
 
@@ -21,7 +21,7 @@ export const leading_zero_bits_64: InstructionRun = (r, args, regs) => {
 
 // LEADING_ZERO_BITS_32
 export const leading_zero_bits_32: InstructionRun = (r, args, regs) => {
-  regs[reg(args.b)] = clz<u32>(u32(regs[reg(args.a)]));
+  regs[reg(args.b)] = u64(clz<u32>(u32(regs[reg(args.a)])));
   return ok(r);
 };
 
@@ -33,7 +33,7 @@ export const trailing_zero_bits_64: InstructionRun = (r, args, regs) => {
 
 // TRAILING_ZERO_BITS_32
 export const trailing_zero_bits_32: InstructionRun = (r, args, regs) => {
-  regs[reg(args.b)] = ctz<u32>(u32(regs[reg(args.a)]));
+  regs[reg(args.b)] = u64(ctz<u32>(u32(regs[reg(args.a)])));
   return ok(r);
 };
